refactoring: move setting classes and interfaces to its own namespaces
