+++ app/Contracts/Settings/OptionsServiceInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Contracts;
+namespace App\Contracts\Settings;
 
 interface OptionsServiceInterface
 {
