+++ app/Providers/AppServiceProvider.php
@@ -4,15 +4,15 @@ namespace App\Providers;
 
 use Illuminate\Contracts\Auth\PasswordBroker as PasswordBrokerContract;
 use App\Services\Auth\AuthService;
-use App\Services\DbOptionsService;
+use App\Services\Settings\DbOptionsService;
 use Illuminate\Support\ServiceProvider;
-use App\Contracts\OptionsServiceInterface;
+use App\Contracts\Settings\OptionsServiceInterface;
 use App\Contracts\Auth\AuthServiceInterface;
-use App\Contracts\SettingsServiceInterface;
+use App\Contracts\Settings\SettingsServiceInterface;
 use App\Contracts\Users\AdminUserServiceInterface;
 use App\Contracts\Users\UserExporterInterface;
 use App\Contracts\Users\UserServiceInterface;
-use App\Services\SettingsService;
+use App\Services\Settings\SettingsService;
 use App\Services\Users\AdminUserService;
 use App\Services\Users\Exporters\CsvUserExporter;
 use App\Services\Users\UserService;
